+++ src/pages/postCreatePage/assets/icons/ic_check_theme.png
