+++ src/pages/postCreatePage/assets/icons/ic_plus_darkGray.png
