+++ src/pages/postCreatePage/constants/images.d.ts
