+++ src/pages/postCreatePage/constants/propTypes.ts
