--- src/pages/postTheme/constants/enum.ts
@@ -1,6 +1,0 @@
-export enum BackgroundColor {
-  Beige = "beige",
-  Purple = "purple",
-  Blue = "blue",
-  Green = "green",
-}
